Are you saying that this linux can run on a computer without windows underneath it, at all ? As in, without a boot disk, without any drivers, and without any services ?
That sounds preposterous to me.

If it were true (and I doubt it), then companies would be selling computers
without a windows. This clearly is not happening, so there must be some error
in your calculations. I hope you realise that windows is more than just Office
? Its a whole system that runs the computer from start to finish, and that is
a very difficult thing to acheive. A lot of people dont realise this.

Microsoft just spent $9 billion and many years to create Vista, so it does not
sound reasonable that some new alternative could just snap into existence
overnight like that. It would take billions of dollars and a massive effort to
achieve. IBM tried, and spent a huge amount of money developing OS/2 but could
never keep up with Windows. Apple tried to create their own system for years,
but finally gave up recently and moved to Intel and Microsoft.

Its just not possible that a freeware like the Linux could be extended to the
point where it runs the entire computer fron start to finish, without using
some of the more critical parts of windows. Not possible.

I think you need to re-examine your assumptions.

    -- jerryleecooper
    -- Talkback on ZDNet (  )