Restraint
Manage a great nation as you would cook a delicate fish.
To govern men in accord with nature
It is best to be restrained;
Restraint makes agreement easy to attain,
And easy agreement builds harmonious relationships;
With sufficient harmony no resistance will arise;
When no resistance arises, then you possess the heart of the
nation,
And when you possess the nation's heart, your influence will long
endure:
Deeply rooted and firmly established.
This is the method of far sight and long life.
		-- Lao Tse, "Tao Te Ching"